requirements.txt
streamlit
pandas
requests
python-dotenv
beautifulsoup4
fpdf2
arabic-reshaper
python-bidi
PyGitHub
duckduckgo-search
streamlit-autorefresh
groq
tavily-python
networkx
Pillow
validators
dnspython
